+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Capture HTTP client request and response.
+
 ## [0.10.0] - 2021-05-07
 ### Added
 - [#90] Capture HTTP response headers in django and flask.